--- .eslintrc.js
@@ -1,14 +1,0 @@
-module.exports = {
-  root: true,
-  extends: [
-    'eslint-config-digitalbazaar',
-    'eslint-config-digitalbazaar/jsdoc'
-  ],
-  env: {
-    node: true
-  },
-  rules: {
-    'jsdoc/check-examples': 'off'
-  },
-  ignorePatterns: ['dist/']
-};
