+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # did-io ChangeLog
 
+## 2.0.0 - 2022-xx-xx
+
+### Changed
+- **BREAKING**: Convert to module (ESM).
+- **BREAKING**: Require Node.js >=14.
+- Update dependencies.
+- Lint module.
+
 ## 1.1.0 - 2021-11-30
 
 ### Added
