+++ README.md
@@ -44,7 +44,7 @@ details.
 
 ## Install
 
-Requires Node.js 12+
+Requires Node.js 14+
 
 To install locally (for development):
