+++ lib/index.js
@@ -1,8 +1,13 @@
 /*!
  * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+export {
+  approvesMethodFor,
+  findVerificationMethod,
+  initKeys,
+  parseDid
+} from './did-io.js';
 
-// translate `main.js` to CommonJS
-require = require('esm')(module);
-module.exports = require('./main.js');
+export {CachedResolver} from './CachedResolver.js';
+
+export {VERIFICATION_RELATIONSHIPS} from './constants.js';
