--- lib/main.js
@@ -1,13 +1,0 @@
-/*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
- */
-export {
-  approvesMethodFor,
-  findVerificationMethod,
-  initKeys,
-  parseDid
-} from './did-io.js';
-
-export {CachedResolver} from './CachedResolver.js';
-
-export {VERIFICATION_RELATIONSHIPS} from './constants.js';
