+++ package.json
@@ -2,11 +2,11 @@
   "name": "@digitalbazaar/did-io",
   "version": "1.1.1-0",
   "description": "A library for managing DIDs (Decentralized Identifiers) and associated data.",
-  "homepage": "http://github.com/digitalbazaar/did-io",
+  "homepage": "https://github.com/digitalbazaar/did-io",
   "author": {
     "name": "Digital Bazaar, Inc.",
     "email": "support@digitalbazaar.com",
-    "url": "http://digitalbazaar.com/"
+    "url": "https://digitalbazaar.com/"
   },
   "contributors": [
     {
@@ -28,47 +28,41 @@
   ],
   "repository": {
     "type": "git",
-    "url": "http://github.com/digitalbazaar/did-io"
+    "url": "https://github.com/digitalbazaar/did-io"
   },
   "bugs": {
     "url": "https://github.com/digitalbazaar/did-io/issues",
     "email": "support@digitalbazaar.com"
   },
   "license": "BSD-3-Clause",
-  "main": "lib/index.js",
-  "module": "lib/main.js",
+  "type": "module",
+  "exports": "./lib/index.js",
   "files": [
     "lib/*.js"
   ],
   "dependencies": {
-    "@digitalbazaar/lru-memoize": "^2.0.0",
-    "esm": "^3.2.25"
+    "@digitalbazaar/lru-memoize": "^3.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.11.0",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/runtime": "^7.11.0",
-    "babel-loader": "^8.1.0",
-    "chai": "^4.2.0",
-    "cross-env": "^7.0.2",
-    "eslint": "^7.22.0",
-    "eslint-config-digitalbazaar": "^2.6.1",
-    "eslint-plugin-jsdoc": "^32.2.0",
-    "karma": "^5.1.1",
-    "karma-babel-preprocessor": "^8.0.1",
+    "c8": "^7.11.2",
+    "chai": "^4.3.6",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.14.0",
+    "eslint-config-digitalbazaar": "^3.0.0",
+    "eslint-plugin-jsdoc": "^39.2.9",
+    "eslint-plugin-unicorn": "^42.0.0",
+    "karma": "^6.3.19",
     "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^3.1.0",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^4.0.2",
-    "mocha": "^8.1.0",
-    "webpack": "^4.44.1"
+    "karma-sourcemap-loader": "^0.3.8",
+    "karma-webpack": "^5.0.0",
+    "mocha": "^9.2.2",
+    "webpack": "^5.72.0"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   },
   "keywords": [
     "Decentralized",
@@ -77,15 +71,12 @@
   ],
   "scripts": {
     "test": "npm run lint && npm run test-node && npm run test-karma",
-    "test-node": "cross-env NODE_ENV=test mocha -r esm --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} test/**/*.spec.js",
-    "test-karma": "karma start karma.conf.js",
+    "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} test/**/*.spec.js",
+    "test-karma": "karma start karma.conf.cjs",
+    "coverage": "cross-env NODE_ENV=test c8 --reporter=lcov --reporter=text-summary npm run test-node",
+    "coverage-ci": "cross-env NODE_ENV=test c8 --reporter=lcovonly --reporter=text npm run test-node",
+    "coverage-report": "c8 report",
     "lint": "eslint .",
     "preversion": "npm test"
-  },
-  "eslintConfig": {
-    "env": {
-      "mocha": true,
-      "node": true
-    }
   }
 }
